ci: fold eval link into the template artifact comment
Drop the separate Eval PR Comment workflow and the section-merge helper;
the pr-template-artifact job now posts a single comment containing both the
evals-monitor link and the run-anywhere template command. One writer, one
comment, plain marker-based upsert — no cross-workflow coordination.

Tradeoff: the eval link now rides the ~5-min build and is skipped on
docs-only PRs (where the artifact job doesn't run and evals are moot),
instead of posting instantly on every PR open.
