+++ .github/workflows/ci.yml
@@ -233,13 +233,12 @@ jobs:
           name: appkit-template-${{ steps.version.outputs.version }}
           path: appkit-template-${{ steps.version.outputs.version }}.zip
 
-      # Build a self-contained, run-anywhere command that downloads this run's
-      # artifact and scaffolds an app from it. Every value (run id, repo,
-      # artifact name, output dir) is baked in so the reader can paste it into
-      # any folder — no repo checkout or specific cwd required. This is the
-      # `template` section of the shared PR-bot comment; the section markers are
-      # added by upsert-pr-comment-section.cjs, not here.
-      - name: Build "try this template" comment section
+      # Build the PR-bot comment: a link to the evals-monitor app, plus a
+      # self-contained, run-anywhere command that downloads this run's artifact
+      # and scaffolds an app from it. Every value (run id, repo, artifact name,
+      # output dir) is baked in so the reader can paste it into any folder — no
+      # repo checkout or specific cwd required.
+      - name: Build PR-bot comment
         env:
           VERSION: ${{ steps.version.outputs.version }}
           RUN_ID: ${{ github.run_id }}
@@ -248,7 +247,11 @@ jobs:
         run: |
           ARTIFACT="appkit-template-${VERSION}"
           OUT="appkit-pr-${PR_NUMBER}"
+          EVALS_URL="https://evals-monitor-6051921418418893.staging.aws.databricksapps.com/prs/appkit/${PR_NUMBER}"
           {
+            echo "<!-- pr-bot -->"
+            echo "> 🔬 &nbsp;**Run evals on this PR** &nbsp;·&nbsp; [**Go to Evals Monitor →**](${EVALS_URL})"
+            echo ""
             echo "### 📦 Try this PR's app template"
             echo ""
             echo "Scaffolds a new app from this PR's SDK build. Run it in **any** folder (requires the GitHub CLI — \`gh auth login\` — and the Databricks CLI):"
@@ -260,20 +263,19 @@ jobs:
             echo '```'
             echo ""
             echo "The template pins \`@databricks/appkit\` and \`@databricks/appkit-ui\` to tarballs built from this branch, so the scaffolded app runs against this PR's code."
-          } > template-section.md
+          } > pr-bot-comment.md
 
       # Fork PRs get a read-only GITHUB_TOKEN, so commenting would 403 and fail
       # a job the author can't fix. Skip for forks — the command is still in the
-      # job log above. Matches bundle-size.yml / eval-pr-comment.yml.
-      - name: Upsert template section into PR comment
+      # job log above. Matches bundle-size.yml.
+      - name: Upsert PR-bot comment
         if: github.event.pull_request.head.repo.full_name == github.repository
         uses: actions/github-script@3a2844b7e9c422d3c10d287c895573f7108da1b3 # v9.0.0
         env:
-          SECTION_ID: template
-          SECTION_PATH: template-section.md
+          COMMENT_PATH: pr-bot-comment.md
         with:
           script: |
-            const upsert = require('./.github/scripts/upsert-pr-comment-section.cjs');
+            const upsert = require('./.github/scripts/upsert-pr-bot-comment.cjs');
             await upsert({ github, context });
 
   docs-build:
