--- .github/workflows/eval-pr-comment.yml
@@ -1,51 +1,0 @@
-name: Eval PR Comment
-
-# Contributes the `evals` section (a link to the evals-monitor app, where the
-# author can start an eval for their PR) to the shared PR-bot comment. We post a
-# link rather than triggering the job from CI because GitHub Actions runners
-# cannot reach the dogfood.staging workspace (network perimeter).
-on:
-  pull_request:
-    types: [opened, reopened]
-    branches: [main]
-
-permissions:
-  contents: read
-  pull-requests: write
-
-# Serialize per-PR so a rapid open+reopen can't race the section upsert.
-# cancel-in-progress: false — the queued run still executes and re-writes an
-# identical section, which is harmless.
-concurrency:
-  group: eval-pr-comment-${{ github.event.pull_request.number }}
-  cancel-in-progress: false
-
-jobs:
-  post-eval-link:
-    name: Post eval link comment
-    # Skip fork PRs: their GITHUB_TOKEN is read-only (createComment would 403),
-    # and the evals-monitor app is internal — the link is useless to external
-    # contributors anyway.
-    if: github.event.pull_request.head.repo.full_name == github.repository
-    runs-on:
-      group: databricks-protected-runner-group
-      labels: linux-ubuntu-latest
-    steps:
-      # Checkout so the github-script step can require the shared section helper.
-      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
-      - name: Build eval link comment section
-        env:
-          PR_NUMBER: ${{ github.event.pull_request.number }}
-        run: |
-          URL="https://evals-monitor-6051921418418893.staging.aws.databricksapps.com/prs/appkit/${PR_NUMBER}"
-          echo "> 🔬 &nbsp;**Run evals on this PR** &nbsp;·&nbsp; [**Go to Evals Monitor →**](${URL})" > eval-section.md
-
-      - name: Upsert eval section into PR comment
-        uses: actions/github-script@3a2844b7e9c422d3c10d287c895573f7108da1b3 # v9.0.0
-        env:
-          SECTION_ID: evals
-          SECTION_PATH: eval-section.md
-        with:
-          script: |
-            const upsert = require('./.github/scripts/upsert-pr-comment-section.cjs');
-            await upsert({ github, context });
